Removing the frontend that test hoce
