--- app/public/css/nivell-1.css
@@ -1,23 +1,0 @@
-#json{
-    display: block;
-    width: 40em;
-    margin:0 auto;
-}
-
-.link{
-    color: #f00;
-    text-decoration: underline;
-}
-
-.link:hover{
-    color: #00f;
-}
-
-#jsonView{
-    color: #ffc107;
-    text-decoration: underline;
-}
-
-#jsonView:hover{
-    color: #f00
-}
