--- app/public/css/style.css
@@ -1,123 +1,0 @@
-*{
-    padding: 0;
-    margin: 0;
-    text-decoration: none;
-    user-select: none;
-}
-
-@font-face {
-    font-family: 'Rowdies';
-    src: url('../assets/fonts/Rowdies-Regular.ttf');
-}
-
-#nav{
-    display:flex;
-    position: fixed;
-    z-index: 100;
-    background-color: #fff;
-    font-family: 'Rowdies';
-    width: 100vw;
-    height: 4em;
-    border-bottom: 1px solid black;
-}
-
-#nav-img{
-    margin: .5em 0 .5em 2em;
-    width:12em;
-}
-
-#nav > ul > a :hover{
-    transition: all .5s;
-    color: #f0f;
-}
-
-#nav > ul{
-    display:flex;
-    align-items: flex-end;
-    justify-content: flex-end;
-    margin-right: 2em;
-    width: 100vw;
-}
-
-#main > h1 {
-    padding-top: 2em;
-}
-
-.links{
-    display: inline-block;
-    font-size: 2em;
-    line-height: 2em;
-    margin-right: 1em;
-}
-
-.e{
-    margin-top: 2em;
-}
-
-.text1{
-    color: #d7137f;
-}
-
-.title{
-    font-size: 3em;
-    font-family: 'Rowdies';
-    text-align: center;
-    width:100vw;
-}
-
-.paragraph{
-    font-size:1.5em;
-    font-family: Arial, Helvetica, sans-serif;
-    margin: 2em 4em;
-    line-height: 2em;
-    text-align: justify;
-}
-
-#footer{
-    position: relative;
-    font-family: Arial, Helvetica, sans-serif;
-    width: 100vw;
-    height:3em;
-    border-top: 1px solid #000;
-    margin-top: 2em;
-}
-
-#footer > p{
-    text-align:center;
-    font-size: 1.5em;
-    line-height: 2em;
-}
-
-#name-link{
-    color:#555;
-}
-
-#name-link:hover{
-    color:#000;
-    text-decoration: underline;
-}
-
-.resolt{
-    display: block;
-    margin:0 auto;
-        position: relative;
-    z-index: 100;
-    font-family: 'Helvetica', Arial, sans-serif;
-    font-size: 1.3em;
-    margin-top: 2em;
-    height:3em;
-    width:12em;
-    background-color: #d7137f;
-    color: #fff;
-    border: none;
-}
-
-.resolt:hover{
-    background-color: #909;
-    border-bottom:3px solid #d7137f;
-    cursor: pointer;
-}
-
-.resolt:active{
-    color: #000;
-}
