--- app/public/css/upload.css
@@ -1,64 +1,0 @@
-#main{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width:100vw;
-    height:100vh;
-    background-image: url('../assets/images/edificios-ciudad-de-noche_2560x1440_3183.webp');
-    background-size: cover;
-    background-attachment: fixed;
-}
-
-#form-container{
-    display: block;
-    width:80vw;
-    height:80vh;
-    margin-top: 5em;
-    background-color: #fff7;
-    box-shadow: 1px 1px 1px #d7137f77, 1px 1px 1px 1px #d7137f77 inset;
-}
-
-#sendImg{
-    display: flex;
-    position: relative;
-    z-index: 20;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-#sendImg > input {
-    display: block;
-    position: relative;
-    z-index: 100;
-    padding:3em;
-    background-color: #0006;
-    font-family: 'Helvetica', Arial, sans-serif;
-    text-shadow: 1px 1px 1px #000;
-    color: #d7137f;
-    font-size: 2em;
-}
-
-#sendImg > button{
-    position: relative;
-    z-index: 100;
-    font-family: 'Helvetica', Arial, sans-serif;
-    font-size: 1.3em;
-    margin-top: 2em;
-    height:3em;
-    width:8em;
-    background-color: #d7137f;
-    color: #fff;
-    border: none;
-}
-
-#sendImg > button:hover{
-    background-color: #909;
-    border-bottom:3px solid #d7137f;
-    cursor: pointer;
-}
-
-#sendImg > button:active{
-    color: #000
-}
