--- app/public/html/nivell-1.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Nivell 1</title>
-    <link rel="icon" href='../assets/images/ITico.ico'>
-    <link rel='stylesheet' type='text/css' href='../css/style.css'>
-    <link rel='stylesheet' type='text/css' href='../css/nivell-1.css'>
-</head>
-
-<body>
-    <nav id='nav'>
-        <a href='/'><img src='./assets/images/IT logo.png' id='nav-img'></a>
-        <ul>
-            <a href='/nivell-1'>
-                <li class='text1 links'>Nivell 1 ⭐️</li>
-            </a>
-            <a href='/nivell-2'>
-                <li class='text1 links'>Nivell 2 ⭐️⭐️</li>
-            </a>
-            <a href='/nivell-3'>
-                <li class='text1 links'>Nivell 3 ⭐️⭐️⭐️</li>
-            </a>
-        </ul>
-    </nav>
-    <main id='main'>
-        <h1 class='text1 title'>🧑🏻‍💻 Entrega 4.1: Node REST Server 🧑🏻‍💻</h1>
-        <h2 class='text1 title e'>📌 Exercici 1 📌</h2>
-        <p class='paragraph'>Crea un servidor amb Express que retorni a una petició GET a l'endpoint <a class='link' href='/user' target='_BLANK'>/user</a> un json amb el teu nom, edat i la url des d'on es fa la petició.</p>
-        <p class='paragraph'>Nota: Per poder veure un fitxer JSON a la majoria de navegadors cal tenir instal·lat una extensió del tipus <a id='jsonView' href='https://chrome.google.com/webstore/detail/jsonview/gmegofmjomhknnokphhckolhcffdaihd?hl=es' target='_BLANK'>JSONView</a></p>
-        <img id='json' src='../assets/images/json.png'>
-        <a href='/user' target='_BLANK'><button class='resolt'>Veure exercici resolt</button></a>
-        <h2 class='text1 title e'>📌 Exercici 2 📌</h2>
-        <p class='paragraph'>Afegeix un endpoint <a class='link' href='/upload'>/upload</a> per a pujar al servidor un arxiu de tipus png, jpg o gif que retorni un missatge d'error en
-        cas que l'extensió de l'arxiu no coincideixi amb aquestes.</p>
-        <a href='/upload'><button class='resolt'>Probar l'exercici</button></a>
-    </main>
-    <footer id='footer'>
-        <p>Creat per <a id='name-link' target='_BLANK' href='http://daniel-espanadero.com/'>Daniel Españadero
-                Fernández</a>
-        </p>
-    </footer>
-</body>
-
-</html>
