--- app/public/html/nivell-2.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Nivell 2</title>
-    <link rel="icon" href='../assets/images/ITico.ico'>
-    <link rel='stylesheet' type='text/css' href='../css/style.css'>
-</head>
-
-<body>
-    <nav id='nav'>
-        <a href='/'><img src='./assets/images/IT logo.png' id='nav-img'></a>
-        <ul>
-            <a href='/nivell-1'>
-                <li class='text1 links'>Nivell 1 ⭐️</li>
-            </a>
-            <a href='/nivell-2'>
-                <li class='text1 links'>Nivell 2 ⭐️⭐️</li>
-            </a>
-            <a href='/nivell-3'>
-                <li class='text1 links'>Nivell 3 ⭐️⭐️⭐️</li>
-            </a>
-        </ul>
-    </nav>
-    <main id='main'>
-        <h1 class='text1 title'>🧑🏻‍💻 Entrega 4.1: Node REST Server 🧑🏻‍💻</h1>
-        <h2 class='text1 title e'>📌 Exercici 1 📌</h2>
-        <p class='paragraph'>Creu un endpoint /time que rebi per POST com a paràmetre un JSON amb el nom d'usuari i retorni un objecte JSON que
-        contingui l'hora i data actual. Inclogui un middleware que afegeixi la capçalera Cache-control: no-cache. Habiliti CORS
-        (Cross-Origin Resource Sharing) en les respostes, ja sigui mitjançant Express o mitjançant un altre middleware.</p>
-        <a href='/time' target='_BLANK'><button class='resolt'>Veure exercici resolt</button></a>
-    </main>
-    <footer id='footer'>
-        <p>Creat per <a id='name-link' target='_BLANK' href='http://daniel-espanadero.com/'>Daniel Españadero
-                Fernández</a>
-        </p>
-    </footer>
-</body>
-
-</html>
