--- app/public/html/nivell-3.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Nivell 3</title>
-    <link rel="icon" href='../assets/images/ITico.ico'>
-    <link rel='stylesheet' type='text/css' href='../css/style.css'>
-</head>
-
-<body>
-    <nav id='nav'>
-        <a href='/'><img src='./assets/images/IT logo.png' id='nav-img'></a>
-        <ul>
-            <a href='/nivell-1'>
-                <li class='text1 links'>Nivell 1 ⭐️</li>
-            </a>
-            <a href='/nivell-2'>
-                <li class='text1 links'>Nivell 2 ⭐️⭐️</li>
-            </a>
-            <a href='/nivell-3'>
-                <li class='text1 links'>Nivell 3 ⭐️⭐️⭐️</li>
-            </a>
-        </ul>
-    </nav>
-    <main id='main'>
-        <h1 class='text1 title'>🧑🏻‍💻 Entrega 4.1: Node REST Server 🧑🏻‍💻</h1>
-        <h2 class='text1 title e'>📌 Exercici 1 📌</h2>
-        <p class='paragraph'>Afegeixi un middleware a l'endpoint anterior que retorni un HTTP Status 401 - Unauthorized si la capçalera de la petició
-        no conté autenticació bàsica (usuari i contrasenya).</p>
-        <a href='/time' target='_BLANK'><button class='resolt'>Veure exercici resolt</button></a>
-    </main>
-    <footer id='footer'>
-        <p>Creat per <a id='name-link' target='_BLANK' href='http://daniel-espanadero.com/'>Daniel Españadero
-                Fernández</a>
-        </p>
-    </footer>
-</body>
-
-</html>
