--- app/public/html/uploaded.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Pujat</title>
-    <link rel='stylesheet' type='text/css' href='../css/style.css'>
-    <link rel='stylesheet' type='text/css' href='../css/uploaded.css'>
-</head>
-<body>
-    <main id='main'>
-    <h1>L'arxiu s'ha pujat correctament</h1>
-    <image src='../assets/images/cohete.png'></image>
-    <a href='/upload'><button class='resolt'>Tornar</button></a>
-    </main>
-</body>
-</html>
