--- app/public/index.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang='es'>
-
-<head>
-    <meta charset='UTF-8'>
-    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
-    <meta name='viewport' content='width=device-width, initial-scale=1.0'>
-    <title>Entrega 4.1: Node REST Server</title>
-    <link rel="icon" href='../assets/images/ITico.ico'>
-    <link rel='stylesheet' type='text/css' href='css/style.css'>
-    <link rel='stylesheet' type='text/css' href='css/index.css'>
-</head>
-
-<body>
-    <nav id='nav'>
-        <a href='/'><img src='./assets/images/IT logo.png' id='nav-img'></a>
-        <ul>
-            <a href='/nivell-1'>
-                <li class='text1 links'>Nivell 1 ⭐️</li>
-            </a>
-            <a href='/nivell-2'>
-                <li class='text1 links'>Nivell 2 ⭐️⭐️</li>
-            </a>
-            <a href='/nivell-3'>
-                <li class='text1 links'>Nivell 3 ⭐️⭐️⭐️</li>
-            </a>
-        </ul>
-    </nav>
-    <main id='main'>
-        <h1 class='text1 title'>🧑🏻‍💻 Entrega 4.1: Node REST Server 🧑🏻‍💻</h1>
-        <img id='node' src='./assets/images/nodejs-2451250.png'>
-        <h2 class='text1 title'>Instruccions del projecte</h2>
-        <p class='paragraph'>Resolució del lliurament 4.1 titulat - Node REST Server - on creo una api rest de resposta ràpida. A més del que demana
-        l'exercici he afegit front-end per aprendre a integrar-lo en un servidor d'express i entendre l'estructura de la millor
-        manera possible, així doncs, hi ha dues maneres de comprovar aquests exercicis; la primera seria a través del navegador
-        web i la segona, seria utilitzant postman.</p>
-        <a href='https://www.postman.com/downloads/' target='_BLANK'><img id='postman' src='./assets/images/0_Ij4wyJ4yMq_0Vm_U.png'></a>
-        <h3 id='estructura' class='text1 title'>🏛 Estructura 🏛</h3>
-        <p class='paragraph'>La manera que utilitzo per crear i organitzar el servidor és express basat en classes, ja que considero que és més
-        senzill d'escalar i mantenir el codi relativament ordenat.
-        A la carpeta <span id='models'>models</span> és on es troba la lògica del servidor i els <span id='controllers'>controllers</span>, les <span id='routes'>routes</span> i els <span id='middlewares'>middlewares</span> els guardo a
-        les seves carpetes corresponents.</p>
-    </main>
-    <footer id='footer'>
-        <p>Creat per <a id='name-link' target='_BLANK' href='http://daniel-espanadero.com/'>Daniel Españadero Fernández</a></p>
-    </footer>
-
-</body>
-
-</html>
