--- app/routes/index.js
@@ -1,9 +1,0 @@
-const { Router } = require('express');
-const {indexGet} = require('../controllers/index');
-
-const router = Router();
-
-router.get('/', indexGet);
-
-
-module.exports = router;
