+++ package.json
@@ -1,21 +1,29 @@
 {
-  "name": "4.1 node REST server",
+  "name": "Entrega-4.1-Node-REST-Server",
   "version": "1.0.0",
-  "description": "",
+  "description": "4.1",
   "main": "index.js",
   "scripts": {
-    "start": "node app/app",
-    "dev": "nodemon app/app",
+    "start": "node src/app",
+    "dev": "nodemon src/app",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "Daniel Españadero Fernández",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DanielEspanadero/nodeInitialDemo.git"
+  },
+  "keywords": [],
+  "author": "",
   "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/DanielEspanadero/nodeInitialDemo/issues"
+  },
+  "homepage": "https://github.com/DanielEspanadero/nodeInitialDemo#readme",
   "dependencies": {
     "cors": "^2.8.5",
-    "dotenv": "^9.0.2",
+    "dotenv": "^16.0.0",
     "express": "^4.17.2",
-    "multer": "^1.4.4",
-    "node": "^17.4.0"
+    "multer": "^1.4.4"
   },
   "devDependencies": {
     "nodemon": "^2.0.15"
