+++ src/controllers/upload.js
@@ -1,0 +1,7 @@
+const uploadPost = (req, res) => {
+    res.json({
+        msg: 'uploaded'
+    });
+};
+
+module.exports = { uploadPost };
